fix(OMN-9888): rename to contract_parser_event_bus.py per naming convention
Files in src/omnibase_core/contracts/ must use the contract_ prefix
(see omnibase_core CLAUDE.md naming table; existing examples:
contract_hash_registry.py, contract_diff_computer.py, contract_loader.py).

Renames src/omnibase_core/contracts/parser_event_bus.py to
contract_parser_event_bus.py and updates the test import. All other surfaces
(EventBusContractShapeError, ModelEventBusSubscription, ModelEventBusParseResult,
test cases) unchanged.

Local re-verification:
- uv run pytest tests/unit/contracts/test_parser_event_bus_shapes.py -v
  -> 7 passed
- uv run mypy --strict on all 4 source files -> 0 errors
- pre-commit run --files (renamed module + test) -> all hooks green

Addresses CodeRabbit major finding on PR #944.
